+++ src/Pim/Bundle/EnrichBundle/Tests/Unit/Form/Subscriber/AddAttributeRequirementsSubscriberTest.php
@@ -3,6 +3,7 @@
 namespace Pim\Bundle\EnrichBundle\Tests\Unit\Form\Subscriber;
 
 use Pim\Bundle\CatalogBundle\Entity\Family;
+use Pim\Bundle\CatalogBundle\Model\AttributeRequirementInterface;
 use Pim\Bundle\EnrichBundle\Form\Subscriber\AddAttributeRequirementsSubscriber;
 
 /**
@@ -114,7 +115,7 @@ class AddAttributeRequirementsSubscriberTest extends \PHPUnit_Framework_TestCase
      * @param mixed $attribute
      * @param mixed $channel
      *
-     * @return \Pim\Bundle\CatalogBundle\Entity\AttributeRequirement
+     * @return AttributeRequirementInterface
      */
     private function getAttributeRequirementMock($attribute, $channel = null)
     {